Clean

+++ base/flags.cpp
@@ -22,10 +22,6 @@ std::unordered_map<string, int*>& flags_int() {
   return internal;
 };
 
-//std::unordered_map<string, string> flag_descriptions_;
-//std::unordered_map<string, string*> flags_string_;
-//std::unordered_map<string, int*> flags_int_;
-
 void ProcessFlags(int* argc, char** argv[]) {
   int newargc = 1;
   std::vector<string> arguments;
@@ -33,7 +29,7 @@ void ProcessFlags(int* argc, char** argv[]) {
   for(int i = 1; i < *argc; i++) {
     string temp((*argv)[i]);
     if (temp.size() > 2 && temp[0] == '-' && temp[1] == '-' && temp[2] != '-' && temp.find('=') != temp.npos) { // it's a flag!
-      int equal_sign_position = temp.find('=');
+      unsigned long equal_sign_position = temp.find('=');
       string name = temp.substr(2, equal_sign_position - 2);
       string value = temp.substr(equal_sign_position + 1);
       if (flags_string().find(name) != flags_string().end()) { // flag is a string
